Выполнение задачи по "Поддержим животных" версия 2

+++ src/main/java/school/faang/sprint4/bjs2_63071/Donation.java
@@ -3,10 +3,12 @@ package school.faang.sprint4.bjs2_63071;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Getter
 public class Donation {
     private final int id;
-    private final double amount;
+    private final BigDecimal amount;
 
 }
